requirements.txt
requests
strands-agents
strands-agents-tools
